requirements.txt
matplotlib
monai-weekly[gdown, nibabel, tqdm, ignite]
numpy
scipy
nibabel
torch